pulling turbine out, starting to remove globals.
