+++ turbofan/engine_validation.py
@@ -10,6 +10,7 @@ import numpy as np
 # Import engine components and maps
 from maps import (FanMap, FanMapPerformance, HPCMap, HPCMapPerformance,
                   LPCMap, LPCMapPerformance)
+from turbine import Turbine, TurbinePerformance
 
 # Import substitution files
 from get_737800_subs import get_737800_subs
@@ -44,7 +45,7 @@ class Engine(Model):
         self.setvals(eng, BLI)
         self.compressor = Compressor()
         self.combustor = Combustor()
-        self.turbine = Turbine()
+        self.turbine = Turbine(LPTeta=LPTeta, lptexp1=lptexp1)
         self.fanmap = FanMap()
         self.lpcmap = LPCMap()
         self.hpcmap = HPCMap()
@@ -806,7 +807,7 @@ class CompressorPerformance(Model):
         diffuser = [
             #free stream stagnation values
              #https://www.grc.nasa.gov/www/k-12/airplane/isentrop.html
-            Tt0 == state["T_{atm}"] / (c1) ** (-1),             #https://www.grc.nasa.gov/www/k-12/airplane/isentrop.html
+            Tt0 == state["T_{atm}"] / (c1) ** (-1),
             ht0 == self.comp['C_{p_{air}'] * Tt0,
 
             #diffuser exit stagnation values (station 1.8)
@@ -911,8 +912,8 @@ class CombustorPerformance(Model):
         T41 = Variable('T_{4.1}', 'K', 'Static Temperature at the Turbine Inlet (4.1)')
 
         #------------------------Variables for cooling flow model---------------------------
-        #define the f plus one variable, limits the number of signomials
-        #variables for station 4a
+        # define the (f+1) variable, limits the number of signomials
+        # for station 4a
         u4a = Variable('u_{4a}', 'm/s', 'Flow Velocity at Station 4a')
         M4a = Variable('M_{4a}', '-', 'User Specified Station 4a Mach #')
         P4a = Variable('P_{4a}', 'kPa', 'Static Pressure at Station 4a (4a)')
@@ -967,77 +968,7 @@ class CombustorPerformance(Model):
 
         return constraints
 
-class Turbine(Model):
-    """"
-    Turbine model
-    """
-    def setup(self):
-        #define new variables
-        #turbines
-        Cpt1 = Variable('C_{p_{t1}}', 1280, 'J/kg/K', "Cp Value for Combustion Products in HP Turbine") #1300K gamma = 1.318
-        Cpt2 = Variable('C_{p_{t2}}', 1184, 'J/kg/K', "Cp Value for Combustion Products in LP Turbine") #800K gamma = 1.354
 
-        #-------------------------diffuser pressure ratios--------------------------
-        pitn = Variable('\\pi_{tn}', '-', 'Turbine Nozzle Pressure Ratio')
-
-        #---------------------------efficiencies & takeoffs-----------------------
-        etaHPshaft = Variable('\eta_{HPshaft}', '-', 'Power Transmission Efficiency of High Pressure Shaft, Smears in Losses for Electrical Power')
-        etaLPshaft = Variable('\eta_{LPshaft}', '-', 'Power Transmission Efficiency of Low Pressure Shaft, Smeras in Losses for Electrical Power')
-
-    def dynamic(self, engine):
-        """
-        creates an instance of the fan map performance model
-        """
-        return TurbinePerformance(self, engine)
-
-class TurbinePerformance(Model):
-    """
-    combustor perfomrance constraints
-    """
-    def setup(self, turbine, engine):
-        self.turbine = turbine
-        self.engine = engine
-
-        #define new variables
-        #------------------LPT inlet stagnation states (station 4.5)------------------
-        ht45 = Variable('h_{t_{4.5}}', 'J/kg', 'Stagnation Enthalpy at the HPT Exit (4.5)')
-        Pt45 = Variable('P_{t_{4.5}}', 'kPa', 'Stagnation Pressure at the HPT Exit (4.5)')
-        Tt45 = Variable('T_{t_{4.5}}', 'K', 'Stagnation Temperature at the HPT Exit (4.5)')
-
-        #-------------------HPT exit (station 4.5) stagnation states--------------------
-        Pt49 = Variable('P_{t_{4.9}}', 'kPa', 'Stagnation Pressure at the HPTExit (49)')
-        Tt49 = Variable('T_{t_{4.9}}', 'K', 'Stagnation Temperature at the HPT Exit (49)')
-        ht49 = Variable('h_{t_{4.9}}', 'J/kg', 'Stagnation Enthalpy at the HPT Exit (49)')
-
-        #------------------------turbo machinery pressure ratios--------------
-        pihpt = Variable('\\pi_{HPT}', '-', 'HPT Pressure Ratio')
-        pilpt = Variable('\\pi_{LPT}', '-', 'LPT Pressure Ratio')
-
-        #------------------turbine nozzle exit stagnation states (station 5)------------
-        Pt5 = Variable('P_{t_5}', 'kPa', 'Stagnation Pressure at the Turbine Nozzle Exit (5)')
-        Tt5 = Variable('T_{t_5}', 'K', 'Stagnation Temperature at the Turbine Nozzle Exit (5)')
-        ht5 = Variable('h_{t_5}', 'J/kg', 'Stagnation Enthalpy at the Turbine Nozzle Exit (5)')
-
-        #make the constraints
-        constraints = []
-
-        #turbine constraints
-        constraints.extend([
-            #HPT Exit states (station 4.5)
-            ht45 == self.turbine['C_{p_{t1}}'] * Tt45,
-
-            #LPT Exit States
-            Pt49 == pilpt * Pt45,
-            pilpt == (Tt49/Tt45)**(lptexp1),    #turbine efficiency is 0.9
-            ht49 == self.turbine['C_{p_{t2}}'] * Tt49,
-
-            #turbine nozzle exit states
-            Pt5 == self.turbine['\\pi_{tn}'] * Pt49, #B.167
-            Tt5 == Tt49,    #B.168
-            ht5 == ht49     #B.169
-            ])
-
-        return constraints
 
 class Thrust(Model):
     """"
